--- Assets/Scripts/Runtime/1.Domain/InGame/Mission/FailCondition/AndFaliConditionGroup.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8094cdd5d4f919740a4f1cdea3111be6
